aider: Import the `slugify` function from the `slugify` module in the `downloader.py` file.
# Aider chat conversation:

USER: I ran this command:

 python main.py "https://www.youtube.com/watch?v=MsEmOQeBcRA"

And got this output:

Processing video: https://www.youtube.com/watch?v=MsEmOQeBcRA
Traceback (most recent call last):
  File "/home/chris/code/summarybot-16k/main.py", line 20, in <module>
    main()
  File "/home/chris/code/summarybot-16k/main.py", line 15, in main
    summary = summarize_video(video_url)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/chris/code/summarybot-16k/summarizer.py", line 22, in summarize_video
    transcript_path = transcribe_video(video_url)
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/chris/code/summarybot-16k/transcriber.py", line 44, in transcribe_video
    audio_file = download_and_check_audio(video_url)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/chris/code/summarybot-16k/downloader.py", line 25, in download_and_check_audio
    audio_file = download_youtube_audio(url)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/chris/code/summarybot-16k/downloader.py", line 10, in download_youtube_audio
    filename = slugify(video.title) + '.webm'
               ^^^^^^^
NameError: name 'slugify' is not defined

ASSISTANT: The error message indicates that the `slugify` function is not defined in the `download_youtube_audio` function in the `downloader.py` file.

The `slugify` function is used to convert the video title into a format that can be used as a filename. It seems that the `slugify` function is not imported in the `downloader.py` file.

To fix this, we need to import the `slugify` function in the `downloader.py` file.

Here is the change that needs to be made:

```python
downloader.py
<<<<<<< HEAD
from pytube import YouTube
from utils import get_video_title
import os
=======
from pytube import YouTube
from utils import get_video_title
from slugify import slugify
import os
>>>>>>> updated
```

This change imports the `slugify` function from the `slugify` module at the top of the `downloader.py` file.